+++ src/main/MainScene.ts
@@ -1,6 +1,7 @@
 import { AudioManager } from "./AudioManager";
 import { InteractiveManager } from "./InteractiveManager";
 import { BackgroundManager } from "./BackgroundManager";
+import { MainEventsManager } from "./MainEventsManager";
 
 /**
  * Main is the gameplay scene which calls the managers to control different aspects of the game
@@ -23,6 +24,8 @@ export class Main extends Phaser.Scene {
         new BackgroundManager(this);
         this.interactiveManager = new InteractiveManager(this);
         this.scene.run("UI");
+        // this.scene.start("Credits");
+        this.scale.on("resize", this.resize);
     }
 
     /**
@@ -30,5 +33,15 @@ export class Main extends Phaser.Scene {
      */
     public update(): void {
         this.interactiveManager.update();
+    }
+
+    /**
+     * Resize gets called when the screen is resized. It fires off an event for the other
+     * game objects to respond to
+     *
+     * @param gameSize - the new size of the screen
+     */
+    private resize(gameSize: Phaser.Structs.Size): void {
+        MainEventsManager.emit("resize", gameSize);
     }
 }