+++ src/preloader/PreloaderScene.ts
@@ -1,5 +1,6 @@
 import { LoadingBar } from "./LoadingBar";
 import * as WebFontLoader from "webfontloader";
+import { GameSettings } from "../utilities/GameSettings";
 
 /**
  * Preloader scene loads all the assets for the main game
@@ -72,10 +73,13 @@ export class Preloader extends Phaser.Scene {
         }
     }
 
+    /**
+     * Loads the webfonts.
+     */
     private loadFonts(): void {
         WebFontLoader.load({
             google: {
-                families: ["Chewy"],
+                families: [GameSettings.DISPLAY_FONT],
             },
         });
     }