+++ src/utilities/GameSettings.ts
@@ -1,0 +1,3 @@
+export class GameSettings {
+    public static DISPLAY_FONT = "Chewy";
+}